Refactor game into modules and add bullet weapon
